Add preprocessing module
